% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{saps-package}
\alias{saps-package}
\title{Implements Significance Analysis of Prognostic Signatures (SAPS), a
robust method for determining prognostically significant gene sets}
\description{
\code{\link{saps}} will usually be the only function needed.
}
\references{
Beck AH, Knoblauch NW, Hefti MM, Kaplan J, Schnitt SJ, et al.
(2013) Significance Analysis of Prognostic Signatures. PLoS Comput Biol 9(1):
e1002875.doi:10.1371/journal.pcbi.1002875
}

