% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculatePEnrichment}
\alias{calculatePEnrichment}
\title{Compute P_enrichment}
\usage{
calculatePEnrichment(rankedGenes, candidateGeneSet, cpus)
}
\arguments{
\item{rankedGenes}{An \emph{nx1} matrix of concordance indices for \emph{n}
genes. Generally this will be the z-score returned by
\code{\link{rankConcordance}}. The row names should contain gene identifiers.}

\item{candidateGeneSet}{A \emph{1xp} matrix of \emph{p} gene identifiers.
The row name should contain a name for the gene set.}

\item{cpus}{This value is passed to the \code{\link[piano]{runGSA}} function
in the \pkg{piano} package. For multi-core CPUs, this value should be set to
the number of cores (which will significantly improve the computational time).}
}
\value{
The function returns a matrix with the following columns:

\code{P_enrichment  direction}

\code{P_enrichment} is the enrichment score, while \code{direction} is
either 1 or -1 depending on the direction of association.
}
\description{
This function performs a pre-ranked gene set enrichment
analysis (GSEA) to evaluate the degree to which a candidate gene set is
overrepresented at the top or bottom extremes of a ranked list of
concordance indices. This function is normally called by
\code{\link{saps}}.
}
\references{
Beck AH, Knoblauch NW, Hefti MM, Kaplan J, Schnitt SJ, et al.
(2013) Significance Analysis of Prognostic Signatures. PLoS Comput Biol 9(1):
e1002875.doi:10.1371/journal.pcbi.1002875

Subramanian A, Tamayo P, Mootha VK, Mukherjee S, Ebert BL, et al. (2005)
Gene set enrichment analysis: a knowledge-based approach for interpreting
genome-wide expression profiles. Proc Natl Acad Sci U S A 102: 15545–15550.
}
\seealso{
\code{\link{saps}} \code{\link[piano]{runGSA}}
}

