% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculatePPure}
\alias{calculatePPure}
\title{Compute P_pure}
\usage{
calculatePPure(geneData, survivalTimes, events)
}
\arguments{
\item{geneData}{An \emph{nxp} matrix consisting of \emph{n} patients
and \emph{p} genes in the candidate prognostic set.}

\item{survivalTimes}{A vector of survival times. The length must equal
the number of rows \emph{n} in \code{geneData}.}

\item{events}{A vector of events. The length must equal the number of
rows \emph{n} in \code{geneData}.}
}
\value{
A log-rank p-value indicating the probability that the two groups
show no survival difference.
}
\description{
This function stratifies patients into two groups via k-means
clustering (k=2) on an \emph{nxp} matrix consisting of \emph{n} patients
and \emph{p} genes in the candidate prognostic set. It is normally called
by \code{\link{saps}}.
}
\references{
Beck AH, Knoblauch NW, Hefti MM, Kaplan J, Schnitt SJ, et al.
(2013) Significance Analysis of Prognostic Signatures. PLoS Comput Biol 9(1):
e1002875.doi:10.1371/journal.pcbi.1002875
}
\seealso{
\code{\link{saps}}
}

