% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculatePRandom}
\alias{calculatePRandom}
\title{Compute P_random}
\usage{
calculatePRandom(dataSet, sampleSize, p_pure, survivalTimes, events,
  random.samples = 25)
}
\arguments{
\item{dataSet}{A matrix, where the column names are gene identifiers
and the values are gene expression levels. Each row should contain data for a
single patient.}

\item{sampleSize}{The desired size for the randomly sampled gene sets.}

\item{p_pure}{The candidate P_pure against which to compare the P_pure
values for the randomly generated gene sets.}

\item{survivalTimes}{A vector of survival times. The length must equal
the number of rows in \code{dataSet}.}

\item{events}{A vector of events. The length must equal the number of
rows in \code{dataSet}.}

\item{random.samples}{The number of random gene sets to sample.}
}
\value{
The proportion of randomly sampled gene sets with a calculated
P_pure at least as significant as the provided \code{p_pure}.
}
\description{
This function randomly samples gene sets, and calculates
P_pure (via \code{\link{calculatePPure}}) for each one. P_random is the
proportion of randomly sampled gene sets achieving a P_pure at least as
significant as the provided \code{p_pure}. This function is normally called
by \code{\link{saps}}.
}
\references{
Beck AH, Knoblauch NW, Hefti MM, Kaplan J, Schnitt SJ, et al.
(2013) Significance Analysis of Prognostic Signatures. PLoS Comput Biol 9(1):
e1002875.doi:10.1371/journal.pcbi.1002875
}
\seealso{
\code{\link{saps}}
}

