% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rankConcordance}
\alias{rankConcordance}
\title{Compute concordance indices}
\usage{
rankConcordance(dataset, survivalTimes, events)
}
\arguments{
\item{dataset}{A matrix, where the column names are gene identifiers
and the values are gene expression levels. Each row should contain data for
a single patient.}

\item{survivalTimes}{A vector of survival times. The length must equal the number of
rows (i.e. patients) in \code{dataset}.}

\item{events}{A vector of events. The length must equal the number of rows
(i.e. patients) in \code{dataset}.}
}
\value{
The function returns a matrix with two columns:

\code{cindex  z}

and as many rows as \code{dataset}. \code{cindex} and \code{z}
are the concordance index and corresponding z-score. The
row names contain the gene identifiers.
}
\description{
Computes concordance indices for a gene expression data set,
and returns the concordance index and the z-score.
}
\details{
This function is a wrapper for
\code{\link[survcomp]{concordance.index}} in the \pkg{survcomp}
package. It applies the latter over the columns of \code{dataset} to
calculate concordance indices and the corresponding z-score for each gene.
}
\seealso{
\code{\link{saps}} \code{\link[survcomp]{concordance.index}}
}

