% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saps}
\alias{saps}
\title{Compute SAPS statistics}
\usage{
saps(candidateGeneSets, dataSet, survivalTimes, events, random.samples = 25,
  cpus = 1, verbose = TRUE)
}
\arguments{
\item{candidateGeneSets}{A matrix with at least one row, where each row represents
a gene set, and the column values are gene identifiers. The row names should contain
unique names for the gene sets. The column values may contain \code{NA} values, since
in general gene sets will have differing lengths.}

\item{dataSet}{A matrix, where the column names are gene identifiers
(in the same format as the values in \code{candidateGeneSets})
and the values are gene expression levels. Each row should contain data for a single
patient.}

\item{survivalTimes}{A vector of survival times. The length must equal the number of
rows (i.e. patients) in \code{dataSet}. This parameter is used by
\code{\link[survcomp]{concordance.index}} to compute the concordance index, which
is in turn used by \code{\link{calculatePEnrichment}} to compute P_enrichment.}

\item{events}{A vector of events. The length must equal the number of rows
(i.e. patients) in \code{dataSet}. This parameter is used by
\code{\link[survcomp]{concordance.index}} to compute the concordance index, which
is in turn used by \code{\link{calculatePEnrichment}} to compute P_enrichment.}

\item{random.samples}{An integer that specifies how many random gene sets to sample
when computing P_random. Used by \code{\link{calculatePRandom}}.}

\item{cpus}{An integer that specifies the number of cpus/cores to be used when
calculating P_enrichment. If greater than 1 (the default), the \pkg{snowfall}
package must be installed or an error will occur.}

\item{verbose}{A boolean indicating whether to display status messages during
computation. Defaults to \code{TRUE}.}
}
\value{
The function returns a matrix with the following columns:

\code{Size P_pure  P_random  P_enrichment  direction}

Each row represents a gene set, and the row names contain the gene set names.
\code{Size} is the size of the gene set. \code{P_pure}, \code{P_random}, and
\code{P_enrichment} are the respective SAPS statistics for the gene set, while
\code{direction} is the direction of association for the enrichment score.
}
\description{
This is the main user interface to the \pkg{saps} package, and is
usually the only function needed.
}
\references{
Beck AH, Knoblauch NW, Hefti MM, Kaplan J, Schnitt SJ, et al.
(2013) Significance Analysis of Prognostic Signatures. PLoS Comput Biol 9(1):
e1002875.doi:10.1371/journal.pcbi.1002875
}

